+++ packages/vitest-environment-miniflare/test/fixtures/modules/core.module.spec.js
@@ -17,3 +17,31 @@ test("handles requests with mocked upstream", async () => {
   const res = await fetch(new Request("https://random.mf"));
   expect(await res.text()).toBe("Hello World!");
 });
+
+test("reply callback options has expected data", async () => {
+  const mockAgent = getMiniflareFetchMock();
+  mockAgent.disableNetConnect();
+  const client = mockAgent.get("https://random.mf");
+  client.intercept({ path: "/", method: "POST" }).reply(200, (opts) => {
+    return opts;
+  });
+  const res = await fetch("https://random.mf", {
+    method: "POST",
+    body: JSON.stringify({ foo: "bar" }),
+  });
+  expect(await res.json()).toEqual({
+    body: '{"foo":"bar"}',
+    bodyTimeout: 300000,
+    headers: {
+      "accept-encoding": "br, gzip, deflate",
+      "content-length": "13",
+      "content-type": "text/plain;charset=UTF-8",
+      "mf-loop": "1",
+    },
+    headersTimeout: 300000,
+    maxRedirections: 0,
+    method: "POST",
+    origin: "https://random.mf",
+    path: "/",
+  });
+});